feat(emacs): Add in emacs role
This is a very basic emacs role that adds the ability to install emacs
along with a tiny bit of configuration for emacs, that mainly revolves
around adding in packages and providing configurations. Right now there
still needs to be a bit more work around this as I have other blobs in
my `.emacs`, but it seems to work for me.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>