fix(emacs): install emacs on apt-based systems
This has a few other small things in around checking for missing files,
but the biggest improvement is that on apt-based systems (i.e. Ubuntu),
`use-package` will be installed along with emacs. This avoids an
annoyance when emacs would throw an error the first time it starts on a
system because the `.emacs` used `use-package` to install the packages.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>