feat(emacs-head): support building emacs-head on MacOS
The default `brew install emacs` installs a console only version of
emacs. While `brew cask install emacs` installs a graphical version, but
it doesn't have support for inline images, which makes org-mode a ton
nicer. This allows you to build the head version of emacs by adding in
the following variables to your playbook:

```yaml
build_from_head: true
build_flags: ["HEAD", "with-cocoa", "with-librsvg", "with-imagemagick", "with-no-frame-refocus", "with-mailutils", "with-dbus", "with-modules", "with-xwidgets"]
```

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>