feat(emacs-plus): migrate to emacs-plus
It seems like emacs-plus is the latest way to install Emacs on your
macOS machine. This chunk of code brings in emacs-plus@28 and installs
the application. It doesn't appear like you can get away with
symlinking it to `/Applications`, so it also copies the application
over. In the future, this should check to see the version of Emacs
installed, and if it figures out that it's installed by this script,
go through and overwrite it. But, until then, you'll need to manually
remove the old version when you update Emacs through homebrew, which
is really annoying now that I think about it.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>